fix: 구조 수정 및 주석 삭제
